requirements.txt
paho-mqtt==1.6.1
keras==2.14.0
numpy==1.26.1
openai==1.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
PyAudio==0.2.14
pyhtml2pdf==0.0.7
python-dotenv==1.0.0
selenium==4.15.2
SpeechRecognition==3.10.0
tensorboard==2.14.1
tensorflow==2.14.0
xhtml2pdf==0.2.13
